Remove RequestFilterFunc/ResponseFilterFunc - move to separate files.
